+++ app/src/main/java/com/example/crearrepositorio/common_ui/BaseFragment.kt
@@ -1,8 +1,7 @@
-package com.example.crearrepositorio.ui.fragments
+package com.example.crearrepositorio.common_ui
 
 import androidx.fragment.app.Fragment
 import androidx.viewbinding.ViewBinding
-import com.example.crearrepositorio.databinding.FragmentHomeBinding
 
 abstract class BaseFragment<T : ViewBinding> : Fragment() {
     protected var _binding: T? = null
