--- app/src/main/java/com/example/crearrepositorio/domain/repositories/MovieRepository.kt
@@ -1,8 +1,0 @@
-package com.example.crearrepositorio.domain.repositories
-
-import com.example.crearrepositorio.domain.entities.MovieModel
-import kotlinx.coroutines.flow.Flow
-
-interface MovieRepository {
-    fun createMovies(): Flow<List<MovieModel>>
-}
