+++ app/src/main/java/com/example/crearrepositorio/features/films/domain/MovieRepository.kt
@@ -1,0 +1,7 @@
+package com.example.crearrepositorio.features.films.domain
+
+import kotlinx.coroutines.flow.Flow
+
+interface MovieRepository {
+    fun createMovies(): Flow<List<MovieModel>>
+}
