+++ app/src/main/java/com/example/crearrepositorio/features/home/ui/FragmentHome.kt
@@ -1,24 +1,26 @@
-package com.example.crearrepositorio.ui.fragments
+package com.example.crearrepositorio.features.home.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import com.example.crearrepositorio.databinding.FragmentHomeBinding
-import com.example.crearrepositorio.ui.replaceFragment
-import com.example.crearrepositorio.ui.viewModel.MainViewModel
-import com.example.crearrepositorio.ui.viewModel.UiState
+import com.example.crearrepositorio.features.films.ui.FirstFragment
+import com.example.crearrepositorio.common_ui.BaseFragment
+import com.example.crearrepositorio.features.series.ui.SecondFragment
+import com.example.crearrepositorio.features.actors.ui.ThirdFragment
+import com.example.crearrepositorio.common_ui.replaceFragment
 import kotlinx.coroutines.launch
 
 class FragmentHome : BaseFragment<FragmentHomeBinding>() {
 
     private val binding get() = _binding!!
-    private val viewModel: MainViewModel by activityViewModels()
+    private val viewModel: HomeViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
