+++ app/src/main/res/layout/fragment_first.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.fragments.FirstFragment">
+    tools:context=".features.films.ui.FirstFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycledMovies"
