+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.fragments.FragmentHome">
+    tools:context=".features.home.ui.FragmentHome">
 
     <Button
         android:id="@+id/button"
