+++ app/src/main/res/layout/fragment_second.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragments.SecondFragment">
+    tools:context=".features.series.ui.SecondFragment">
 
     <TextView
         android:layout_width="wrap_content"
